Change logic of call for papers email notification
Call for papers email logic used to have some 'nice' code around it
to ensure it was sent even if the app was offline on the day it
should have been triggered.
But that made other things worse (once) so now just converted to a
simple check of 'is this the day? y/n'.