Hide edit box for semesters
There is a bit of livewire 'funk' going on despite there being
a wire:key on each row on the table which makes it update the
semester on the wrong course - leading to sad times.
So changed to just display the semester for now.
Admins can edit an individual course and change the semester that
way for the moment.